move db to infra layer
